requirements.txt
# To install all dependencies: python -m pip install -r requirements.txt

python-dotenv

# -- NFC functionality --
pyscard

# --CSV functionality--
pandas

# --AES/RSA/AES and RSA--
pycryptodome

# --ECC--
cryptography

# --HILLCIPHER--
numpy

# --DATABASE-- 
pymongo

# Notes:
# - It's best if each algorithm has their own directory in your project,
# since the section of the codes that reads data from the CSV will always 
# assume the CSV is one level up. Additionally it helps in organization of 
# pem and bin files.
